homework7-Shanaurin-Anton/requirements.txt
pytest==7.4.3
pytest-postgresql==5.0.0
psycopg2-binary==2.9.10
faker==20.1.0
python-dotenv==1.0.0